feat: fixing security filter with all roles

+++ src/main/java/com/moreira/loginauthapi/application/services/UserService.java
@@ -3,6 +3,8 @@ package com.moreira.loginauthapi.application.services;
 import com.moreira.loginauthapi.application.dto.*;
 import com.moreira.loginauthapi.domain.entities.Role;
 import com.moreira.loginauthapi.domain.entities.User;
+import com.moreira.loginauthapi.domain.exceptions.EmailInUseException;
+import com.moreira.loginauthapi.domain.exceptions.PasswordException;
 import com.moreira.loginauthapi.domain.repositories.RoleRepository;
 import com.moreira.loginauthapi.domain.repositories.UserRepository;
 import com.moreira.loginauthapi.infra.security.TokenService;
@@ -10,13 +12,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
+import java.nio.file.AccessDeniedException;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -37,21 +41,23 @@ public class UserService {
     public UserRegistrationResponseDTO createUser(UserRegistrationRequestDTO userRegistrationRequestDTO) {
         //antes de tudo, chegar se email já existe no banco de dados, se sim, continua
 
+        if (userRepository.existsByEmail(userRegistrationRequestDTO.email())) {
+            throw new EmailInUseException("Email already in use.");
+        }
+
+        if (userRegistrationRequestDTO.password().length() < 6) {
+            throw new PasswordException("Password must be at least 6 characters.");
+        }
+
         User user = new User();
         //adicionar validações para exceção
         user.setName(userRegistrationRequestDTO.name());
         user.setEmail(userRegistrationRequestDTO.email());
         user.setPassword(passwordEncoder.encode(userRegistrationRequestDTO.password()));
 
+        user.setRole("ROLE_COMMON");
+
         //lembrar de arrumar no SecurityConfig os endpoints com as roles
-
-        //role principal
-        Role primaryRole = roleRepository.findByName(userRegistrationRequestDTO.role());
-        user.setPrimaryRole(primaryRole);
-
-
-        //adiciona role principal na lista de roles
-        user.getRoles().add(primaryRole);
 
         userRepository.save(user);
 
@@ -62,10 +68,10 @@ public class UserService {
                 user.getName(),
                 user.getEmail(),
                 user.getPassword(),
-                user.getRoles(),
                 token
         );
     }
+
 
     public Page<UserPageableResponseDTO> findAllPageable(Pageable pageable) {
         Page<User> users = userRepository.findAll(pageable);
@@ -80,7 +86,7 @@ public class UserService {
     public UserChangesResponseDTO update(String id, UserChangesRequestDto userChangesRequestDto) {
         //implementar metodo para buscar por ID lançando exceção automaticamente
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
 
         user.setName(userChangesRequestDto.name());
         user.setEmail(userChangesRequestDto.email());
@@ -95,7 +101,7 @@ public class UserService {
 
     public UserInformationsResponseDTO findById(String id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
 
         return new UserInformationsResponseDTO(
                 user.getId(),
@@ -105,6 +111,40 @@ public class UserService {
     }
 
     public void deleteById(String id) {
+        if (!userRepository.existsById(id)) {
+            throw new UsernameNotFoundException("User not found.");
+        }
         userRepository.deleteById(id);
+    }
+
+    @Secured(("ROLE_ADMIN"))
+    public UserWithNewRoleResponseDTO insertRoleOnUser(String id, UserWithNewRoleRequestDTO userWithNewRoleRequestDTO) throws AccessDeniedException {
+
+        if (!userRepository.existsById(id)) {
+            throw new UsernameNotFoundException("User not found.");
+        }
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            throw new AccessDeniedException("You dont have the permission to do this action.");
+        }
+
+        User user = userRepository.getReferenceById(id);
+        Set<Role> roles = new HashSet<>(roleRepository.findByNameIn(userWithNewRoleRequestDTO.roles()));
+
+        user.getRoles().addAll(roles);
+
+        userRepository.save(user);
+
+        return new UserWithNewRoleResponseDTO(
+                user.getId(),
+                user.getName(),
+                user.getEmail(),
+                user.getRoles().stream().map(Role::getName).collect(Collectors.toSet())
+        );
     }
 }
